+++ src/__generated__/BidResult_sale_artwork.graphql.ts
@@ -1,0 +1,96 @@
+/* tslint:disable */
+
+import { ConcreteFragment } from "relay-runtime";
+export type BidResult_sale_artwork = {
+    readonly current_bid: ({
+        readonly amount: string | null;
+        readonly cents: number | null;
+        readonly display: string | null;
+    }) | null;
+    readonly sale: ({
+        readonly live_start_at: string | null;
+        readonly end_at: string | null;
+    }) | null;
+};
+
+
+
+const node: ConcreteFragment = (function(){
+var v0 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "__id",
+  "args": null,
+  "storageKey": null
+};
+return {
+  "kind": "Fragment",
+  "name": "BidResult_sale_artwork",
+  "type": "SaleArtwork",
+  "metadata": null,
+  "argumentDefinitions": [],
+  "selections": [
+    {
+      "kind": "LinkedField",
+      "alias": null,
+      "name": "current_bid",
+      "storageKey": null,
+      "args": null,
+      "concreteType": "SaleArtworkCurrentBid",
+      "plural": false,
+      "selections": [
+        {
+          "kind": "ScalarField",
+          "alias": null,
+          "name": "amount",
+          "args": null,
+          "storageKey": null
+        },
+        {
+          "kind": "ScalarField",
+          "alias": null,
+          "name": "cents",
+          "args": null,
+          "storageKey": null
+        },
+        {
+          "kind": "ScalarField",
+          "alias": null,
+          "name": "display",
+          "args": null,
+          "storageKey": null
+        }
+      ]
+    },
+    {
+      "kind": "LinkedField",
+      "alias": null,
+      "name": "sale",
+      "storageKey": null,
+      "args": null,
+      "concreteType": "Sale",
+      "plural": false,
+      "selections": [
+        {
+          "kind": "ScalarField",
+          "alias": null,
+          "name": "live_start_at",
+          "args": null,
+          "storageKey": null
+        },
+        {
+          "kind": "ScalarField",
+          "alias": null,
+          "name": "end_at",
+          "args": null,
+          "storageKey": null
+        },
+        v0
+      ]
+    },
+    v0
+  ]
+};
+})();
+(node as any).hash = '8b4ef9f1b528bcd831d1bb5404834b8a';
+export default node;