+++ src/__generated__/QueryRenderersBidFlowQuery.graphql.ts
@@ -55,6 +55,21 @@ fragment ConfirmBid_sale_artwork on SaleArtwork {
     __id
   }
   lot_label
+  ...BidResult_sale_artwork
+  __id
+}
+
+fragment BidResult_sale_artwork on SaleArtwork {
+  current_bid {
+    amount
+    cents
+    display
+  }
+  sale {
+    live_start_at
+    end_at
+    __id
+  }
   __id
 }
 */
@@ -100,6 +115,20 @@ v3 = {
 v4 = {
   "kind": "ScalarField",
   "alias": null,
+  "name": "display",
+  "args": null,
+  "storageKey": null
+},
+v5 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "cents",
+  "args": null,
+  "storageKey": null
+},
+v6 = {
+  "kind": "ScalarField",
+  "alias": null,
   "name": "id",
   "args": null,
   "storageKey": null
@@ -108,7 +137,7 @@ return {
   "kind": "Request",
   "operationKind": "query",
   "name": "QueryRenderersBidFlowQuery",
-  "id": "0a615ed0d3710d159ec8fc60f04c2de2",
+  "id": "8289114d6d466bbd952484339fa7bf21",
   "text": null,
   "metadata": {},
   "fragment": {
@@ -181,20 +210,8 @@ return {
                 "concreteType": "BidIncrementsFormatted",
                 "plural": true,
                 "selections": [
-                  {
-                    "kind": "ScalarField",
-                    "alias": null,
-                    "name": "display",
-                    "args": null,
-                    "storageKey": null
-                  },
-                  {
-                    "kind": "ScalarField",
-                    "alias": null,
-                    "name": "cents",
-                    "args": null,
-                    "storageKey": null
-                  }
+                  v4,
+                  v5
                 ]
               },
               {
@@ -206,8 +223,22 @@ return {
                 "concreteType": "Sale",
                 "plural": false,
                 "selections": [
-                  v4,
-                  v3
+                  v6,
+                  v3,
+                  {
+                    "kind": "ScalarField",
+                    "alias": null,
+                    "name": "live_start_at",
+                    "args": null,
+                    "storageKey": null
+                  },
+                  {
+                    "kind": "ScalarField",
+                    "alias": null,
+                    "name": "end_at",
+                    "args": null,
+                    "storageKey": null
+                  }
                 ]
               },
               {
@@ -219,7 +250,7 @@ return {
                 "concreteType": "Artwork",
                 "plural": false,
                 "selections": [
-                  v4,
+                  v6,
                   {
                     "kind": "ScalarField",
                     "alias": null,
@@ -250,6 +281,26 @@ return {
                 "name": "lot_label",
                 "args": null,
                 "storageKey": null
+              },
+              {
+                "kind": "LinkedField",
+                "alias": null,
+                "name": "current_bid",
+                "storageKey": null,
+                "args": null,
+                "concreteType": "SaleArtworkCurrentBid",
+                "plural": false,
+                "selections": [
+                  {
+                    "kind": "ScalarField",
+                    "alias": null,
+                    "name": "amount",
+                    "args": null,
+                    "storageKey": null
+                  },
+                  v5,
+                  v4
+                ]
               },
               v3
             ]