+++ src/lib/Components/Bidding/Screens/ConfirmBid.tsx
@@ -23,7 +23,7 @@ import { Title } from "../Components/Title"
 import SwitchBoard from "lib/NativeModules/SwitchBoard"
 import { metaphysics } from "../../../metaphysics"
 
-import { BidResult } from "./BidResult"
+import { BidResultScreen } from "./BidResult"
 
 import { ConfirmBid_sale_artwork } from "__generated__/ConfirmBid_sale_artwork.graphql"
 import { Checkbox } from "../Components/Checkbox"
@@ -97,6 +97,7 @@ export class ConfirmBid extends React.Component<ConfirmBidProps, ConformBidState
   verifyBidPosition(results, errors) {
     const status = results.createBidderPosition.result.status
     if (!errors && status === "SUCCESS") {
+      // bidder position is created but not processed yet
       const positionId = results.createBidderPosition.result.position.id
       const query = `
         {
@@ -109,13 +110,17 @@ export class ConfirmBid extends React.Component<ConfirmBidProps, ConformBidState
         }
       `
       const interval = setInterval(() => {
-        metaphysics({ query }).then(this.checkBidPosition.bind(this))
+        metaphysics({
+          query,
+        }).then(this.checkBidPosition.bind(this))
       }, 1000)
-      this.setState({ intervalToken: interval })
+      this.setState({
+        intervalToken: interval,
+      })
     } else {
       const message_header = results.createBidderPosition.result.message_header
       const message_description_md = results.createBidderPosition.result.message_description_md
-      this.showBidResult(false, message_header, message_description_md)
+      this.showBidResult(false, status, message_header, message_description_md)
     }
   }
 
@@ -126,10 +131,11 @@ export class ConfirmBid extends React.Component<ConfirmBidProps, ConformBidState
       clearInterval(this.state.intervalToken)
       if (bidderPosition.is_active) {
         // wining
-        this.showBidResult(true)
+        this.showBidResult(true, "SUCCESS")
       } else {
         // outbid
-        this.showBidResult(false)
+        // TODO: show message (get from metaphysics?)
+        this.showBidResult(false, "ERROR_BID_LOW")
       }
     } else {
       // poll again
@@ -140,11 +146,13 @@ export class ConfirmBid extends React.Component<ConfirmBidProps, ConformBidState
     }
   }
 
-  showBidResult(winning: boolean, messageHeader?: string, messageDescriptionMd?: string) {
+  showBidResult(winning: boolean, status: string, messageHeader?: string, messageDescriptionMd?: string) {
     this.props.navigator.push({
-      component: BidResult,
+      component: BidResultScreen,
       title: "",
       passProps: {
+        sale_artwork: this.props.sale_artwork,
+        status,
         message_header: messageHeader,
         message_description_md: messageDescriptionMd,
         winning,
@@ -226,6 +234,7 @@ export const ConfirmBidScreen = createFragmentContainer(
         artist_names
       }
       lot_label
+      ...BidResult_sale_artwork
     }
   `
 )