+++ src/lib/Components/Bidding/__stories__/BidFlow.story.tsx
@@ -14,7 +14,7 @@ import { Checkbox } from "../Components/Checkbox"
 
 import BidFlow from "../../../Containers/BidFlow"
 import { Input } from "../Components/Input"
-import { BidResult } from "../Screens/BidResult"
+import { BidResultScreen } from "../Screens/BidResult"
 import { ConfirmBidScreen } from "../Screens/ConfirmBid"
 import { ConfirmFirstTimeBid } from "../Screens/ConfirmFirstTimeBid"
 import { MaxBidScreen } from "../Screens/SelectMaxBid"
@@ -23,40 +23,32 @@ const testSaleArtworkID = "5ae73b417622dd026f0fe473"
 const testArtworkID = "ran-hwang-ephemeral-blossom-pp"
 const testSaleID = "cityarts-benefit-auction-2018"
 
-const SelectMaxBidRenderer: React.SFC<any> = ({ render, saleArtworkID }) => {
-  return (
-    <QueryRenderer
-      environment={createEnvironment()}
-      query={graphql`
-        query BidFlowSelectMaxBidRendererQuery($saleArtworkID: String!) {
-          sale_artwork(id: $saleArtworkID) {
-            ...SelectMaxBid_sale_artwork
-          }
-        }
-      `}
-      variables={{
-        saleArtworkID,
-      }}
-      render={render}
-    />
-  )
-}
+const selectMaxBidQuery = graphql`
+  query BidFlowSelectMaxBidRendererQuery($saleArtworkID: String!) {
+    sale_artwork(id: $saleArtworkID) {
+      ...SelectMaxBid_sale_artwork
+    }
+  }
+`
 
-const ConfirmBidScreenRenderer: React.SFC<any> = ({ render, saleArtworkID }) => {
-  return (
-    <QueryRenderer
-      environment={createEnvironment()}
-      query={graphql`
-        query BidFlowConfirmBidScreenRendererQuery($saleArtworkID: String!) {
-          sale_artwork(id: $saleArtworkID) {
-            ...ConfirmBid_sale_artwork
-          }
-        }
-      `}
-      variables={{ saleArtworkID }}
-      render={render}
-    />
-  )
+const confirmBidQuery = graphql`
+  query BidFlowConfirmBidScreenRendererQuery($saleArtworkID: String!) {
+    sale_artwork(id: $saleArtworkID) {
+      ...ConfirmBid_sale_artwork
+    }
+  }
+`
+
+const bidResultQuery = graphql`
+  query BidFlowBidResultScreenRendererQuery($saleArtworkID: String!) {
+    sale_artwork(id: $saleArtworkID) {
+      ...BidResult_sale_artwork
+    }
+  }
+`
+
+const BidFlowStoryRenderer: React.SFC<any> = ({ render, query, saleArtworkID }) => {
+  return <QueryRenderer environment={createEnvironment()} query={query} variables={{ saleArtworkID }} render={render} />
 }
 
 storiesOf("Bidding")
@@ -64,14 +56,19 @@ storiesOf("Bidding")
     return <BidFlowRenderer render={renderWithLoadProgress(BidFlow)} artworkID={testArtworkID} saleID={testSaleID} />
   })
   .add("Select Max Bid", () => (
-    <SelectMaxBidRenderer render={renderWithLoadProgress(MaxBidScreen)} saleArtworkID={testSaleArtworkID} />
+    <BidFlowStoryRenderer
+      render={renderWithLoadProgress(MaxBidScreen)}
+      query={selectMaxBidQuery}
+      saleArtworkID={testSaleArtworkID}
+    />
   ))
   .add("Confirm Bid", () => {
     return (
-      <ConfirmBidScreenRenderer
+      <BidFlowStoryRenderer
         render={renderWithLoadProgress(ConfirmBidScreen, {
           bid: { display: "$20,000", cents: 2000000 },
         })}
+        query={confirmBidQuery}
         saleArtworkID={testSaleArtworkID}
       />
     )
@@ -96,10 +93,53 @@ storiesOf("Bidding")
     )
   })
   .add("Bidding Result (winning)", () => {
-    return <BidResult winning />
+    const status = "SUCCESS"
+    return (
+      <BidFlowStoryRenderer
+        render={renderWithLoadProgress(BidResultScreen, {
+          winning: true,
+          status,
+        })}
+        query={bidResultQuery}
+        saleArtworkID={testSaleArtworkID}
+      />
+    )
   })
   .add("Bidding Result (not highest bid)", () => {
-    return <BidResult winning={false} />
+    const status = "ERROR_BID_LOW"
+    const messageHeader = "Your bid wasn’t high enough"
+    const messageDescriptionMd = `Another bidder placed a higher max bid or the same max bid before you did.  \
+ Bid again to take the lead.`
+    return (
+      <BidFlowStoryRenderer
+        render={renderWithLoadProgress(BidResultScreen, {
+          winning: false,
+          status,
+          message_header: messageHeader,
+          message_description_md: messageDescriptionMd,
+        })}
+        query={bidResultQuery}
+        saleArtworkID={testSaleArtworkID}
+      />
+    )
+  })
+  .add("Bidding Result (live bidding started)", () => {
+    const status = "ERROR_LIVE_BIDDING_STARTED"
+    const messageHeader = "Live bidding has started"
+    const messageDescriptionMd = `Sorry, your bid wasn’t received before live bidding started.\
+ To continue bidding, please [join the live auction](http://live-staging.artsy.net/).`
+    return (
+      <BidFlowStoryRenderer
+        render={renderWithLoadProgress(BidResultScreen, {
+          winning: false,
+          status,
+          message_header: messageHeader,
+          message_description_md: messageDescriptionMd,
+        })}
+        query={bidResultQuery}
+        saleArtworkID={testSaleArtworkID}
+      />
+    )
   })
 
 storiesOf("App Style/Input")