+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 * Adds a `<Input>` component - yuki24
 * Adds a billing address screen - yuki24
 * Reflect style updates for the `<Checkbox>` component - yuki24
+* Load Real values into `BidResult` screen + reimplement `<MarkdownRenderer>` - sepans
 
 ### 1.5.0